+++ fsictools.py
@@ -13,6 +13,7 @@ import re
 from typing import Dict, List
 
 from fsic import BaseModel, Symbol
+import fsic
 
 
 def symbols_to_dataframe(symbols: List[Symbol]) -> 'pandas.DataFrame':
@@ -20,6 +21,28 @@ def symbols_to_dataframe(symbols: List[Symbol]) -> 'pandas.DataFrame':
     from pandas import DataFrame
 
     return DataFrame([s._asdict() for s in symbols])
+
+def symbols_to_graph(symbols: List[Symbol]) -> 'networkx.DiGraph':
+    """Convert the list of symbols to a NetworkX DiGraph. **Requires `networkx`."""
+    import networkx as nx
+
+    G = nx.DiGraph()
+
+    equations = [s.equation for s in symbols if s.equation is not None]
+    for e in equations:
+        lhs, rhs = e.split('=', maxsplit=1)
+        endogenous = [m.group(0) for m in fsic.term_re.finditer(lhs)]
+        exogenous =  [m.group(0) for m in fsic.term_re.finditer(rhs)]
+
+        # Add the equations as node properties
+        G.add_nodes_from(endogenous, equation=e)
+
+        # Add the edges
+        for n in endogenous:
+            for x in exogenous:
+                G.add_edge(x, n)
+
+    return G
 
 def symbols_to_sympy(symbols: List[Symbol]) -> Dict['sympy.Symbol', 'sympy.Eq']:
     """Convert the system of equations into a dictionary of `SymPy` objects. **Requires `SymPy`**."""